% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqp_data.R
\name{aqp_data}
\alias{aqp_data}
\title{Load SoilProfileCollection Datasets}
\usage{
aqp_data(
  ...,
  list = character(),
  verbose = getOption("verbose"),
  envir = .GlobalEnv,
  overwrite = TRUE
)
}
\arguments{
\item{...}{Literal character strings or names.}

\item{list}{character. A vector of object names to load.}

\item{verbose}{logical. If \code{TRUE}, additional diagnostics are printed.}

\item{envir}{The environment where the data should be loaded.}

\item{overwrite}{logical. Should existing objects of the same name in \code{envir} be replaced? Default: \code{TRUE}}
}
\value{
Called for side-effect of loading the specified dataset into global environment.
}
\description{
Loads specified data sets, or list the available data sets.
}
\examples{
\dontshow{if (requireNamespace("aqp", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

aqp_data(loafercreek)
\dontshow{\}) # examplesIf}
}
