% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStaticNASIS.R
\name{createStaticNASIS}
\alias{createStaticNASIS}
\title{Create a memory or file-based instance of NASIS database (for selected
tables)}
\usage{
createStaticNASIS(
  tables = NULL,
  SS = TRUE,
  ignore_pattern = "^sys|_SS|_State|NONDEL|View_0|dm_|xml_",
  static_path = NULL,
  output_path = NULL,
  verbose = FALSE
)
}
\arguments{
\item{tables}{Character vector of target tables. Default: \code{NULL} is all
tables meeting the following criteria.}

\item{SS}{Logical. Include "selected set" tables (ending with suffix
\code{"_View_1"}). Default: \code{TRUE}}

\item{ignore_pattern}{A regular expression identifying tables in NASIS schema to ignore. Default: \code{"^sys|_SS|_State|NONDEL|View_0|dm_"}}

\item{static_path}{Optional: path to SQLite database containing NASIS table
structure; Default: \code{NULL}}

\item{output_path}{Optional: path to new/existing SQLite database to write
tables to. Default: \code{NULL} returns table results as named list.}

\item{verbose}{Issue error messages for unqueryable tables?}
}
\value{
A named list of results from calling \code{dbQueryNASIS} for all
columns in each NASIS table.
}
\description{
Create a memory or file-based instance of NASIS database (for selected
tables)
}
\examples{


\dontrun{
 str(createStaticNASIS(tables = c("calculation","formtext")))
}


}
