% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQueryNASIS.R
\name{dbConnectNASIS}
\alias{dbConnectNASIS}
\title{Create a connection to a local NASIS database}
\usage{
dbConnectNASIS()
}
\value{
A \code{DBIConnection} object, as returned by \code{DBI::dbConnect()}.
}
\description{
Create a connection to a local NASIS database
}
