% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQueryNASIS.R
\name{dbQueryNASIS}
\alias{dbQueryNASIS}
\title{Query a NASIS DBIConnection}
\usage{
dbQueryNASIS(conn, q, close = TRUE, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as returned by \code{DBI::dbConnect()}.}

\item{q}{A statement to execute using \code{DBI::dbGetQuery}; or a (named) vector containing multiple statements to evaluate separately}

\item{close}{Close connection after query? Default: \code{TRUE}}

\item{...}{Additional arguments to \code{DBI::dbGetQuery}}
}
\value{
Result of \code{DBI::dbGetQuery}
}
\description{
Send queries to a NASIS DBIConnection
}
