% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStaticNASIS.R
\name{.dump_NASIS_table}
\alias{.dump_NASIS_table}
\title{Method for "dumping" contents of an entire NASIS table}
\usage{
.dump_NASIS_table(table_name, dsn = NULL)
}
\arguments{
\item{table_name}{Character name of table.}

\item{dsn}{Optional: path to SQLite database containing NASIS table
structure; Default: \code{NULL}}
}
\value{
A data.frame or other result of \code{DBI::dbGetQuery}
}
\description{
Method for "dumping" contents of an entire NASIS table
}
