% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NASIS_table_fkey_by_name.R
\name{.get_NASIS_table_pkey_by_name}
\alias{.get_NASIS_table_pkey_by_name}
\title{Get the key by table name}
\usage{
.get_NASIS_table_pkey_by_name(tables)
}
\arguments{
\item{tables}{character vector of table names}
}
\value{
The foreign key column name for the specified table name
}
\description{
Get the key by table name
}
\examples{
\dontrun{
.get_NASIS_table_pkey_by_name(c("site","phorizon_View_1","not_a_table"))
}
}
