% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateColorMixture.R
\name{estimateColorMixture}
\alias{estimateColorMixture}
\title{Estimate color mixture using L, A, B color coordinates and weights}
\usage{
estimateColorMixture(x, wt = "pct", backTransform = FALSE)
}
\arguments{
\item{x}{data.frame, typically from NASIS containing at least CIE LAB ('L', 'A', 'B') and some kind of weight}

\item{wt}{fractional weights, usually area of hz face}

\item{backTransform}{logical, should the mixed sRGB representation of soil color be transformed to closest Munsell chips? This is performed by aqp::rgb2Munsell default: \code{FALSE}}
}
\value{
A data.frame containing estimated color mixture
}
\description{
Estimate color mixture using L, A, B color coordinates and weights
}
