% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLDM.R
\name{fetchLDM}
\alias{fetchLDM}
\title{Query data from Kellogg Soil Survey Laboratory Data Mart via Soil Data Access}
\usage{
fetchLDM(
  x,
  dsn = NULL,
  what = "pedlabsampnum",
  bycol = "pedon_key",
  tables = c("lab_physical_properties", "lab_mineralogy_glass_count",
    "lab_chemical_properties", "lab_major_and_trace_elements_and_oxides",
    "lab_xray_and_thermal", "lab_calculations_including_estimates_and_default_values",
    "lab_rosetta_Key"),
  chunk.size = 1000,
  ntries = 3
)
}
\arguments{
\item{x}{a vector of values to find in column specified by \code{what}}

\item{dsn}{data source name; either a path to a SQLite database, an open DBIConnection or (default) \code{NULL} (to use \code{soilDB::SDA_query})}

\item{what}{a single column name from tables: \code{lab_combine_nasis_ncss}, \code{lab_webmap}, \code{lab_site}, \code{lab_pedon} or \code{lab_area}}

\item{bycol}{a single column name from \code{lab_layer} used for processing chunks; default: \code{"pedon_key"}}

\item{tables}{a vector of table names; one or more of: \code{"lab_physical_properties"}, \code{"lab_mineralogy_glass_count"}, \code{"lab_chemical_properties"}, \code{"lab_major_and_trace_elements_and_oxides"}, \code{"lab_xray_and_thermal"}, \code{"lab_calculations_including_estimates_and_default_values"}, \code{"lab_rosetta_Key"}}

\item{chunk.size}{number of pedons per chunk (for queries that may exceed \code{maxJsonLength})}

\item{ntries}{number of tries (times to halve \code{chunk.size}) before returning \code{NULL}; default \code{3}}
}
\value{
a \code{SoilProfileCollection}
}
\description{
LDM model diagram:
}
\details{
\itemize{
\item https://jneme910.github.io/Lab_Data_Mart_Documentation/Documents/SDA_KSSL_Data_model.html
}

If the \code{chunk.size} parameter is set too large and the Soil Data Access request fails, the algorithm will attempt to halve the \code{chunk.size} argument up to 3 times, re-trying the query with a smaller chunk.

Currently the \code{lab_area} tables are joined only for the "Soil Survey Area" related records.
}
\examples{
\donttest{

if(requireNamespace("curl") &
   curl::has_internet()) {
   
  # fetch by ssa_key
  fetchLDM(8297, what = "ssa_key")

  # get physical properties for points correlated as taxonomic subgroup "Typic Argialbolls"
  fetchLDM("Typic Argialbolls", what = "corr_taxsubgrp", tables = "lab_physical_properties")
 
  # fetch by area_code (must be a soil survey area code for now; 
  #                      the lowest-level area that is always populated)
  fetchLDM("CA630", what = "area_code")
}

}
}
