% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSTEDUS.R
\name{fetchSTEDUS}
\alias{fetchSTEDUS}
\title{Fetch Standardized Elevation Derivatives of the United States (STEDUS) Grids}
\usage{
fetchSTEDUS(x = NULL, variables = NULL, filename = NULL, overwrite = FALSE)
}
\arguments{
\item{x}{An R spatial object (such as a \emph{SpatVector}, \emph{SpatRaster}, or \emph{sf} object) or a
\emph{SoilProfileCollection} with coordinates initialized via \verb{aqp::initSpatial<-}. Default: \code{NULL}
returns the CONUS extent as virtual raster. If \code{x} is a \emph{SpatRaster} the coordinate reference
system, extent, and resolution are used as a template for the output raster.}

\item{variables}{One or more variables corresponding to STEDUS grid file names (without .tif
extension). See \url{https://storage.googleapis.com/dsm-ft-stedus-30m-public/index.html} for
details.}

\item{filename}{character. Path to write output raster file. Default: \code{NULL} will keep result in
memory (or store in temporary file if memory threshold is exceeded)}

\item{overwrite}{Overwrite \code{filename} if it exists? Default: \code{FALSE}}
}
\value{
A \emph{SpatRaster} object.
}
\description{
This tool creates a virtual raster or downloads data for an extent from Cloud Optimized GeoTIFFs
(COGs) from the Standardized Elevation Derivatives of the United States (STEDUS) project.
}
\examples{
\dontshow{if (curl::has_internet() && requireNamespace("sf") && !inherits(try(requireNamespace("terra", quietly = TRUE), silent = TRUE), 'try-error')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
b <- c(-119.747629, -119.67935, 36.912019, 36.944987)

bbox.sp <- sf::st_as_sf(wk::rct(
  xmin = b[1], xmax = b[2], ymin = b[3], ymax = b[4],
  crs = sf::st_crs(4326)
))

ssurgo.geom <- soilDB::SDA_spatialQuery(
  bbox.sp,
  what = 'mupolygon',
  db = 'SSURGO',
  geomIntersection = TRUE
)

# grid output
res <- fetchSTEDUS(
  ssurgo.geom,
  variables = c("sl_2", "tri_8", "aspct_2", "mca_10000")
)

res

plot(res) 
}
\dontshow{\}) # examplesIf}
}
\references{
Brungard et al. (2024) 30 meter Standardized Elevation Derivatives of the United
States (STEDUS30). Manuscript in preparation. Available online:
https://storage.googleapis.com/dsm-ft-stedus-30m-public/index.html
}
