% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWDB-requests-and-util.R
\name{get_AWDB_elements}
\alias{get_AWDB_elements}
\title{Get AWDB Element (Sensor) Metadata}
\usage{
get_AWDB_elements(...)
}
\arguments{
\item{...}{Additional arguments to \code{.soilDB_curl_get_JSON()}, such as \code{timeout = 120}}
}
\value{
A \code{data.frame} unless network or server error, then \code{NULL}
}
\description{
This function retrieves all element (sensor) metadata from the AWDB webservice.
}
