% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors_from_pedon_db.R
\name{get_colors_from_pedon_db}
\alias{get_colors_from_pedon_db}
\title{Get Soil Color Data from a PedonPC Database}
\usage{
get_colors_from_pedon_db(dsn)
}
\arguments{
\item{dsn}{The path to a 'pedon.mdb' database.}
}
\value{
A data.frame with the results.
}
\description{
Get, format, mix, and return color data from a PedonPC database.
}
\seealso{
\code{\link{get_hz_data_from_pedon_db}},
\code{\link{get_site_data_from_pedon_db}}
}
\author{
Dylan E. Beaudette and Jay M. Skovlin
}
\keyword{manip}
