% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecosite_history_from_NASIS_db.R
\name{get_ecosite_history_from_NASIS_db}
\alias{get_ecosite_history_from_NASIS_db}
\title{Get Site Ecological Site History}
\usage{
get_ecosite_history_from_NASIS_db(best = TRUE, SS = TRUE, dsn = NULL)
}
\arguments{
\item{best}{Should the "best" ecological site correlation be chosen? Creates field called \code{es_selection_method} with \code{"most recent"} or \code{"least missing data"} for resolving many:1 relationships in site history.}

\item{SS}{Use selected set? Default: \code{TRUE}}

\item{dsn}{Path to SQLite data source, or a \code{DBIConnection} to database with NASIS schema.}
}
\value{
a \code{data.frame}, or \code{NULL} on error
}
\description{
Gets the Site Ecological Site History data from local NASIS database. Used by \code{\link[=get_extended_data_from_NASIS_db]{get_extended_data_from_NASIS_db()}}.
}
\seealso{
\code{\link[=get_extended_data_from_NASIS_db]{get_extended_data_from_NASIS_db()}}
}
