% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extended_data_from_pedon_db.R
\name{get_extended_data_from_pedon_db}
\alias{get_extended_data_from_pedon_db}
\title{Get accessory tables and summaries from a local pedonPC Database}
\usage{
get_extended_data_from_pedon_db(dsn)
}
\arguments{
\item{dsn}{The path to a 'pedon.mdb' database.}
}
\value{
A list with the results.
}
\description{
Get accessory tables and summaries from a local pedonPC Database.
}
\seealso{
\code{\link{get_hz_data_from_pedon_db}},
\code{\link{get_site_data_from_pedon_db}}
}
\author{
Jay M. Skovlin and Dylan E. Beaudette
}
\keyword{manip}
