% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hz_data_from_pedon_db.R
\name{get_hz_data_from_pedon_db}
\alias{get_hz_data_from_pedon_db}
\title{Get Horizon Data from a PedonPC Database}
\usage{
get_hz_data_from_pedon_db(dsn)
}
\arguments{
\item{dsn}{The path to a 'pedon.mdb' database.}
}
\value{
A data.frame.
}
\description{
Get horizon-level data from a PedonPC database.
}
\note{
NULL total rock fragment values are assumed to represent an \emph{absence}
of rock fragments, and set to 0.
}
\seealso{
\code{\link{get_colors_from_pedon_db}},
\code{\link{get_site_data_from_pedon_db}}
}
\author{
Dylan E. Beaudette and Jay M. Skovlin
}
\keyword{manip}
