% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lablayer_data_from_NASIS_db.R
\name{get_lablayer_data_from_NASIS_db}
\alias{get_lablayer_data_from_NASIS_db}
\title{Extract lab pedon layer data from a local NASIS Database}
\usage{
get_lablayer_data_from_NASIS_db(SS = TRUE, dsn = NULL)
}
\arguments{
\item{SS}{fetch data from the currently loaded selected set in NASIS or from
the entire local database (default: \code{TRUE})}

\item{dsn}{Optional: path to local SQLite database containing NASIS
table structure; default: \code{NULL}}
}
\value{
A data.frame.
}
\description{
Get lab pedon layer-level (horizon-level) data from a local NASIS database.
}
\note{
This function queries KSSL laboratory site/horizon data from a local
NASIS database from the lab layer data table.
}
\seealso{
\code{\link{get_labpedon_data_from_NASIS_db}}
}
\author{
Jay M. Skovlin and Dylan E. Beaudette
}
\keyword{manip}
