% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NASIS_table_key_by_name.R
\name{get_NASIS_table_key_by_name}
\alias{get_NASIS_table_key_by_name}
\alias{get_NASIS_fkey_by_name}
\alias{get_NASIS_pkeyref_by_name}
\alias{get_NASIS_pkey_by_name}
\title{Get a NASIS table key by type and table name}
\usage{
get_NASIS_table_key_by_name(
  tables,
  keycol = c("all", "fkey", "pkeyref", "pkey")
)
}
\arguments{
\item{tables}{character vector of table names}

\item{keycol}{One of: "fkey" the foreign key; "pkeyref" the primary key referenced by the foreign key, or "pkey" the primary key.}
}
\value{
The key column name for the specified table name
}
\description{
Get a NASIS table key by type and table name
}
\examples{
\dontrun{
get_NASIS_table_key_by_name(c("site","phorizon_View_1","not_a_table"))
}
}
