% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NASISWebReport.R
\name{get_NASISWebReport}
\alias{get_NASISWebReport}
\alias{get_NASISWebProjects}
\title{Low-level methods for getting projects and tables from NASIS Web Reports}
\usage{
get_NASISWebReport(report_name, proj_id, verbose = TRUE)

get_NASISWebProjects(msso, fy, project, verbose = TRUE)
}
\arguments{
\item{report_name}{internal report name}

\item{proj_id}{integer; Project record ID}

\item{verbose}{Default: \code{TRUE}; print messages with report name and parameters?}

\item{msso}{MLRA soil survey office e.g. \code{"2-SON"}}

\item{fy}{Fiscal year e.g. \code{2020}}

\item{project}{Expression to match in project name e.g. \code{"MLRA 18\%"}}
}
\value{
A data.frame, or (invisible) try-error if request fails
}
\description{
\code{get_NASISWebReport()}: Gets a table from the "ReportData" section of specified \code{report_name}

\code{get_NASISWebProjects()}: Gets a table of project name, state, approval statu and project record ID for specified office, fiscal year and project name pattern.
}
