% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchNOAA.R
\name{get_NOAA_GHCND_by_stationyear}
\alias{get_NOAA_GHCND_by_stationyear}
\title{Get Global Historical Climatology Network Daily (GHCND) data from NOAA API for given datatype(s), stationid and year.}
\usage{
get_NOAA_GHCND_by_stationyear(stationid, year, datatypeid, apitoken)
}
\arguments{
\item{stationid}{Station ID (e.g. \code{GHCND:USC00388786})}

\item{year}{A single year (e.g. 2018)}

\item{datatypeid}{One or more NOAA GHCND data type IDs (e.g \code{c("PRCP","SNOW")})}

\item{apitoken}{API key token for NOAA NCDC web services ()}
}
\value{
A data.frame containing the GHCND data requested (limit 1000 records)
}
\description{
Obtain daily summary data for a single station ID and year for one or more datatypes. Note that results from the NOAA API, even with an API token, are limited to 1000 records. Obtaining yearly daily summaries by this method can generally only obtain two data types at a time, for a single year and station, in a single query (e.g. \code{2*365 < 1000}).

  In order to use this function, you must obtain an API token from this website: https://www.ncdc.noaa.gov/cdo-web/token
}
\examples{

#' ## in order to use this function, you must obtain an API token from this website:
##  https://www.ncdc.noaa.gov/cdo-web/token
# get_NOAA_GHCND_by_stationyear(GHCND:USC00388786, year = 2017, datatypeid = "PRCP", apitoken = "yourtokenhere")

}
