% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phroots_from_NASIS_db.R
\name{get_phroots_from_NASIS_db}
\alias{get_phroots_from_NASIS_db}
\title{Get pedon horizon roots data from a local NASIS Database}
\usage{
get_phroots_from_NASIS_db(SS = TRUE, dsn = NULL)
}
\arguments{
\item{SS}{logical, limit query to the selected set}

\item{dsn}{optional path or \emph{DBIConnection} to \link[=NASISLocalDatabase]{local database containing NASIS table structure}; default: \code{NULL}}
}
\value{
A \code{data.frame}
}
\description{
This function returns records from the \code{phroots} table of a local NASIS database. Pedon and pedon horizon record IDs are also included for linking back to related records, typically queried via \code{\link[=fetchNASIS]{fetchNASIS()}}.
}
