% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_from_NASIS.R
\name{get_projectmilestone_from_NASIS}
\alias{get_projectmilestone_from_NASIS}
\alias{get_projectmapunit_from_NASIS}
\title{Get NASIS Project Information}
\usage{
get_projectmilestone_from_NASIS(SS = TRUE, stringsAsFactors = NULL, dsn = NULL)

get_projectmapunit_from_NASIS(SS = TRUE, stringsAsFactors = NULL, dsn = NULL)
}
\arguments{
\item{SS}{\emph{logical}. Use selected set? Default: \code{TRUE}. Set \code{FALSE} for local database.}

\item{stringsAsFactors}{Deprecated.}

\item{dsn}{Optional: path or \emph{DBIConnection} to \link[=NASISLocalDatabase]{local database containing NASIS table structure}; default: \code{NULL}}
}
\value{
\code{get_projectmilestone_from_NASIS()}: \emph{data.frame} containing project and project milestone information

\code{get_projectmapunit_from_NASIS()}: \emph{data.frame} containing project and project mapunit information
}
\description{
Helper functions for accessing the NASIS Project object.
}
