% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_cosurfmorph.R
\name{get_SDA_cosurfmorph}
\alias{get_SDA_cosurfmorph}
\title{Get Geomorphic/Surface Morphometry Data from Soil Data Access}
\usage{
get_SDA_cosurfmorph(
  WHERE = NULL,
  by = "compname",
  table = "cosurfmorphgc",
  dsn = NULL,
  query_string = FALSE
)
}
\arguments{
\item{WHERE}{WHERE clause added to SQL query. For example: \code{areasymbol = 'CA067'}}

\item{by}{Grouping variable. Default: \code{"compname"}}

\item{table}{Target table to summarize. Default: \code{"cosurfmorphgc"} (3D Geomorphic Component). Alternate choices include \code{cosurfmorphhpp} (2D Hillslope Position) and \code{cosurfmorphss} (Surface Shape).}

\item{dsn}{Path to local SSURGO database SQLite database. Default \code{NULL} uses Soil Data Access.}

\item{query_string}{Return query instead of sending to Soil Data Access / local database. Default: \code{FALSE}.}
}
\value{
a \code{data.frame} containing the grouping variable (\code{by}) and tabular summaries of counts and proportions of geomorphic records.
}
\description{
Get Geomorphic/Surface Morphometry Data from Soil Data Access or a local SSURGO data source and summarize by counts and proportions ("probabilities").
}
\details{
Default \code{table="cosurfmorphgc"} summarizes columns \code{geomposmntn}, \code{geomposhill}, \code{geomposflats}, and \code{geompostrce}.
\code{table="cosurfmorphhpp"} summarizes \code{"hillslopeprof"} and \code{table="cosurfmorphss"} summarizes \code{shapeacross} and \code{shapedown}

Queries are a generalization of now-deprecated functions from {sharpshootR} by Dylan Beaudette: \code{geomPosMountainProbability()}, \code{geomPosHillProbability()}, \code{surfaceShapeProbability()}, \code{hillslopeProbability()}

Similar summaries of SSURGO component surface morphometry data by series name can be found in \code{fetchOSD(, extended=TRUE)} or downloaded from \url{https://github.com/ncss-tech/SoilWeb-data}
Full component data including surface morphometry summaries at the "site" level can be obtained with \code{fetchSDA()}.
}
\examples{
\dontrun{
 # Summarize by 3D geomorphic components by component name (default `by='compname'`)
 get_SDA_cosurfmorph("areasymbol = 'CA630'")
 
 # Whole Soil Survey Area summary (using `by = 'areasymbol'`)
 get_SDA_cosurfmorph("areasymbol = 'CA630'", by = 'areasymbol')
 
 # 2D Hillslope Position summary(using `table = 'cosurfmorphhpp'`)
 get_SDA_cosurfmorph("areasymbol = 'CA630'", table = 'cosurfmorphhpp')
 
 # Surface Shape summary (using `table = 'cosurfmorphss'`)
 get_SDA_cosurfmorph("areasymbol = 'CA630'", table = 'cosurfmorphss')
}
}
\seealso{
\code{fetchSDA()} \code{get_SDA_pmgroupname()}
}
\author{
Dylan E. Beaudette, Andrew Brown
}
