% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_hydric.R
\name{get_SDA_hydric}
\alias{get_SDA_hydric}
\title{Get map unit hydric soils information from Soil Data Access}
\usage{
get_SDA_hydric(
  areasymbols = NULL,
  mukeys = NULL,
  method = "MAPUNIT",
  query_string = FALSE
)
}
\arguments{
\item{areasymbols}{vector of soil survey area symbols}

\item{mukeys}{vector of map unit keys}

\item{method}{One of: \code{"Mapunit"}, \code{"Dominant Component"}, \code{"Dominant Condition"}, \code{"None"}}

\item{query_string}{Default: \code{FALSE}; if \code{TRUE} return a character string containing query that would be sent to SDA via \code{SDA_query}}
}
\value{
a data.frame
}
\description{
Assess the hydric soils composition of a map unit.
}
\details{
The default classes for \code{method="MAPUNIT"} are as follows:
\itemize{
\item \code{'Nonhydric'} - no hydric components
\item \code{'Hydric'} - all hydric components
\item \code{'Predominantly Hydric'} - 50\% or more of component proportion is hydric
\item \code{'Partially Hydric'} - some proportion of the MAJOR components are hydric
\item \code{'Predominantly Nonhydric'} - less than 50\% of component proportion is hydric
}

The default result will also include the following summaries of component percentages: \code{total_comppct}, \code{hydric_majors} and \code{hydric_inclusions}.

Default \code{method} \code{"Mapunit"} produces aggregate summaries of all components in the mapunit. Use \code{"Dominant Component"}, \code{"Dominant Condition"} to get the dominant component (highest percentage) or dominant hydric condition (similar conditions aggregated across components). Use \code{"None"} for no aggregation (one record per component).
}
\author{
Jason Nemecek, Chad Ferguson, Andrew Brown
}
