% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_data_from_pedon_db.R
\name{get_site_data_from_pedon_db}
\alias{get_site_data_from_pedon_db}
\title{Get Site Data from a PedonPC Database}
\usage{
get_site_data_from_pedon_db(dsn)
}
\arguments{
\item{dsn}{The path to a 'pedon.mdb' database.}
}
\value{
A data.frame.
}
\description{
Get site-level data from a PedonPC database.
}
\seealso{
\code{\link{get_hz_data_from_pedon_db}},
\code{\link{get_veg_from_AK_Site}},
}
\author{
Dylan E. Beaudette and Jay M. Skovlin
}
\keyword{manip}
