% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soilseries_from_NASIS.R
\name{get_soilseries_from_NASIS}
\alias{get_soilseries_from_NASIS}
\alias{get_soilseries_from_NASISWebReport}
\title{Get records from the Soil Classification (SC) database}
\usage{
get_soilseries_from_NASIS(
  stringsAsFactors = default.stringsAsFactors(),
  static_path = NULL
)
}
\arguments{
\item{stringsAsFactors}{logical: should character vectors be converted to
factors? This argument is passed to the uncode() function. It does not
convert those vectors that have set outside of uncode() (i.e. hard coded).
The 'factory-fresh' default is TRUE, but this can be changed by setting
options(stringsAsFactors = FALSE)}

\item{soils}{character vector of soil series names}
}
\value{
A \code{data.frame}.
}
\description{
These functions return records from the Soil Classification database, either
from the local NASIS database (all series) or via web report (named series
only).
}
\author{
Stephen Roecker
}
\keyword{manip}
