% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSRI.R
\name{get_SRI_layers}
\alias{get_SRI_layers}
\title{Get SRI Layers}
\usage{
get_SRI_layers(gdb)
}
\arguments{
\item{gdb}{A \code{character} of the gdb, e.g. \code{'Deschutes'}.}
}
\value{
A list of metadata about the gdb.
}
\description{
Get SRI Layers
}
\note{
Refer to \code{\link{get_SRI}} for information on gdb availability.
}
\examples{
\donttest{
sri_layers <- get_SRI_layers('Willamette')
}

}
