% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SSURGO.R
\name{get_SSURGO}
\alias{get_SSURGO}
\title{Get SSURGO ZIP files from Web Soil Survey Download Soils Data}
\usage{
get_SSURGO(
  WHERE,
  destdir,
  exdir = destdir,
  extract = TRUE,
  remove_zip = FALSE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{WHERE}{A \code{WHERE} clause expression to be used to filter records in \code{sacatalog} table. Alternately \code{WHERE} can be an \code{sf} or \code{sp} spatial object defining the target extent.}

\item{destdir}{Directory to download ZIP files into.}

\item{exdir}{Directory to extract ZIP archives into. May be a directory that does not yet exist. Each ZIP file will extract to a folder labeled with \code{areasymbol} in this directory. Default: \code{destdir}}

\item{extract}{Logical. Extract ZIP files to \code{exdir}? Default: \code{TRUE}}

\item{remove_zip}{Logical. Remove ZIP files after extracting? Default: \code{FALSE}}

\item{overwrite}{Logical. Overwrite by re-extracting if directory already exists? Default: \code{FALSE}}

\item{quiet}{Logical. Passed to \code{download.file()}.}
}
\value{
character (invisibly) paths to downloaded ZIP files. May not exist if \code{remove_zip = TRUE}.
}
\description{
Download ZIP files containing spatial (ESRI shapefile) and tabular (TXT) files with SSURGO format.
}
\details{
To specify which Soil Survey Areas you would like to obtain data for, specify a \code{WHERE} clause for query of \code{sacatalog} table such as \code{areasymbol = 'CA067'}, \code{"areasymbol IN ('CA628', 'CA067')"} or  \verb{areasymbol LIKE 'CT\%'}.

Pipe-delimited TXT files are found in \emph{/tabular/} folder extracted from a SSURGO ZIP. The files are named for tables in the SSURGO schema. There is no header / the files do not have column names. See the \emph{Soil Data Access Tables and Columns Report}: \url{https://sdmdataaccess.nrcs.usda.gov/documents/TablesAndColumnsReport.pdf} for details on tables, column names and metadata including the default sequence of columns used in TXT files.

Several ESRI shapefiles are found in the \emph{/spatial/} folder extracted from a SSURGO ZIP. These have prefix \code{soilmu_} (mapunit), \code{soilsa_} (survey area), \code{soilsf_} (special features). There will also be a TXT file with prefix \code{soilsf_} describing any special features. Shapefile names then have an \code{a_} (polygon), \code{l_} (line), \code{p_} (point) followed by the soil survey area symbol.
}
