% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_veg_from_AK_Site.R
\name{get_veg_from_AK_Site}
\alias{get_veg_from_AK_Site}
\title{Get Vegetation Data from an AK Site Database}
\usage{
get_veg_from_AK_Site(dsn)
}
\arguments{
\item{dsn}{file path the the AK Site access database}
}
\value{
A data.frame with vegetation data in long format, linked to site ID.
}
\description{
Get Vegetation Data from an AK Site Database
}
\seealso{
\code{\link{get_hz_data_from_pedon_db}},
\code{\link{get_site_data_from_pedon_db}}
}
\author{
Dylan E. Beaudette
}
\keyword{manip}
