% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_veg_other_from_MT_veg_db.R
\name{get_veg_other_from_MT_veg_db}
\alias{get_veg_other_from_MT_veg_db}
\title{Get cover composition data from a Montana RangeDB database}
\usage{
get_veg_other_from_MT_veg_db(dsn)
}
\arguments{
\item{dsn}{The name of the Montana RangeDB front-end database connection
(see details).}
}
\value{
A data.frame.
}
\description{
Get cover composition data from a Montana RangeDB database.
}
\seealso{
\code{\link{get_veg_from_MT_veg_db}},
\code{\link{get_veg_species_from_MT_veg_db}}
}
\author{
Jay M. Skovlin
}
\keyword{manip}
