% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHzErrorsNASIS.R
\name{getHzErrorsNASIS}
\alias{getHzErrorsNASIS}
\title{Check pedon horizon table for logic errors}
\usage{
getHzErrorsNASIS(strict = TRUE)
}
\arguments{
\item{strict}{how strict should horizon boundaries be checked for consistency: TRUE=more | FALSE=less}
}
\value{
A data.frame containing problematic records with columns: 'peiid','pedon_id','hzdept','hzdepb','hzname'
}
\description{
Check pedon horizon table for logic errors
}
