\name{gSSURGO.chunk}
\alias{gSSURGO.chunk}
\docType{data}
\title{Gridded SSURGO Chunk}
\description{A chunk of the gridded SSURGO database (gSSURGO)}
\usage{data(gSSURGO.chunk)}

\details{This is a 106x137 grid, cropped from the gSSURGO database. Cell values are map unit keys (mukey), stored as integers. No raster attribute table (RAT) is included with this sample data set. Note that this sample of the gSSURGO data has been modified such that cell values are map unit keys, rather than the gSSURGO integer key.}

\source{\url{http://soils.usda.gov/survey/geography/ssurgo/description_gssurgo.html}}

\examples{
data(gSSURGO.chunk)

}
\keyword{datasets}
