% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_query.R
\name{makeChunks}
\alias{makeChunks}
\title{Generate chunk labels for splitting data}
\usage{
makeChunks(ids, size = 100)
}
\arguments{
\item{ids}{vector of IDs}

\item{size}{chunk (group) size}
}
\value{
A numeric vector
}
\description{
Generate chunk labels for splitting data
}
\examples{

# split the lowercase alphabet into 2 chunks

aggregate(letters,
          by = list(makeChunks(letters, size=13)),
          FUN = paste0, collapse=",")

}
