% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDB-package.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{NASIS 7 Metadata}
\format{
A \code{data.frame} with the following columns:
\itemize{
\item \code{DomainID} - Integer. ID that uniquely identifies a domain in a data model, not just within a database.
\item \code{DomainName} - Character. Domain Name.
\item \code{DomainRanked} - Integer. Is domain ranked? \code{0} = No; \code{1} = Yes
\item \code{DisplayLabel} - Character. Domain Display Label.
\item \code{ChoiceSequence} - Integer. Order or sequence of Choices.
\item \code{ChoiceValue} - Integer. Value of choice level.
\item \code{ChoiceName} - Character. Name of choice level.
\item \code{ChoiceLabel} - Character. Label of choice level.
\item \code{ChoiceObsolete} - Integer. Is choice level obsolete? \code{0} = No; \code{1} = Yes
\item \code{ColumnPhysicalName} - Character. Physical column name.
\item \code{ColumnLogicalName} - Character. Logical column name.
}
}
\description{
NASIS 7 Metadata from MetadataDomainDetail, MetadataDomainMaster, and MetadataTableColumn tables
}
\keyword{datasets}
