% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_and_clean_colors.R
\name{mix_and_clean_colors}
\alias{mix_and_clean_colors}
\title{Mix and Clean Colors}
\usage{
mix_and_clean_colors(x, wt = "pct", backTransform = FALSE)
}
\arguments{
\item{x}{a \code{data.frame} object containing sRGB coordinates (\code{'r'}, \code{'g'}, \code{'b'}) in [0,1]}

\item{wt}{fractional weights, usually area of hz face}

\item{backTransform}{logical, should the mixed sRGB representation of soil
color be transformed to closest Munsell chips? This is performed by}
}
\value{
A data.frame containing mixed colors
}
\description{
Deprecated: only used in PedonPC functionality; use \code{estimateColorMixture} instead
}
