% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mukey-WCS.R
\name{mukey.wcs}
\alias{mukey.wcs}
\title{Get Map Unit Key (\code{mukey}) grid from SoilWeb Web Coverage Service (WCS)}
\usage{
mukey.wcs(
  aoi,
  db = c("gNATSGO", "gSSURGO", "RSS", "STATSGO", "fSSURGO", "PR_SSURGO", "HI_SSURGO"),
  res = NULL,
  quiet = FALSE
)
}
\arguments{
\item{aoi}{area of interest (AOI) defined using a terra pakcage objects (\code{SpatRaster}, \code{SpatVector}, \code{ext}), sf objects (\code{sf}, \code{sfc}), raster package object (\code{RasterLayer}, \code{bbox}), sp package \verb{Spatial*}, or a \code{list}, see details}

\item{db}{character, mukey grid source, see details}

\item{res}{grid resolution, leave as NULL to use native grid resolution, see details}

\item{quiet}{logical, passed to \code{curl::curl_download()} to enable / suppress URL and progress bar for download.}
}
\value{
A \code{SpatRaster} (or \code{RasterLayer}) object containing indexed map unit keys and associated raster attribute table, or a \code{try-error} if the WCS request fails. Basic metadata are encoded into the resulting \code{SpatRaster}, accessible via \code{terra::metags()}.
}
\description{
Download chunks of the gNATSGO, gSSURGO, OCONUS SSURGO, STATSGO2, and RSS map unit key grid via bounding-box from the SoilWeb WCS.
}
\details{
When specified as a \code{list}, \code{aoi} should contain:
\itemize{
\item \code{aoi}: bounding-box specified as (xmin, ymin, xmax, ymax) e.g. c(-114.16, 47.65, -114.08, 47.68)
\item \code{crs}: coordinate reference system of BBOX, e.g. 'OGC:CRS84' (EPSG:4326, WGS84 Longitude/Latitude)
}

The WCS query is parameterized using a rectangular extent derived from the above AOI specification, after conversion to the native CRS (EPSG:5070) of the WCS grids.

Databases available from this WCS can be queried using \code{WCS_details(wcs = 'mukey')}.\tabular{lll}{
   db \tab crs \tab description \cr
   ak_ssurgo \tab EPSG:3338 \tab AK map unit keys \cr
   as_ssurgo \tab EPSG:4326 \tab AS map unit keys \cr
   fssurgo \tab EPSG:5070 \tab SSURGO/STATSGO2 map unit keys \cr
   gnatsgo \tab EPSG:5070 \tab gNATSGO map unit keys \cr
   gssurgo \tab EPSG:5070 \tab gSSURGO map unit keys \cr
   gu_ssurgo \tab EPSG:4326 \tab GU map unit keys \cr
   hi_ssurgo \tab EPSG:6628 \tab HI map unit keys \cr
   mp_ssurgo \tab EPSG:4326 \tab MP map unit keys \cr
   pr_ssurgo \tab EPSG:32161 \tab PR map unit keys \cr
   pw_ssurgo \tab EPSG:4326 \tab PW map unit keys \cr
   rss \tab EPSG:5070 \tab RSS map unit keys \cr
   statsgo \tab EPSG:5070 \tab STATSGO2 map unit keys \cr
}


The \code{fSSURGO} database is an unofficial hybrid of gSSURGO, back-filled with STATSGO2 data where SSURGO data are missing (e.b. denied access, NOTCOM, large misc. areas other than water).
}
\examples{
\dontrun{
library(terra)

res <- mukey.wcs(list(aoi = c(-116.7400, 35.2904, -116.7072, 35.3026), crs = "EPSG:4326"),
                 db = 'gNATSGO', res = 30) 
  
m <- unique(values(res))

prp <- setNames(
  get_SDA_property(
    c("ph1to1h2o_r", "claytotal_r"),
    "weighted average",
    mukeys = m,
    top_depth = 0,
    bottom_depth = 25,
    include_minors = TRUE,
    miscellaneous_areas = FALSE
  )[, c("mukey", "ph1to1h2o_r", "claytotal_r")],
  c("ID",    "pH1to1_0to25", "clay_0to25")
)

levels(res) <- prp
res2 <- catalyze(res)
res2

plot(res2[['pH1to1_0to25']])
}
}
\author{
D.E. Beaudette and A.G. Brown
}
