% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDB-package.R
\docType{data}
\name{NASIS_table_column_fkey}
\alias{NASIS_table_column_fkey}
\title{NASIS 7 Tables, Columns and Foreign Keys}
\description{
This dataset contains NASIS 7 Tables, Columns and Foreign Keys
}
\keyword{datasets}
