% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncode.R
\name{NASISDomainsAsFactor}
\alias{NASISDomainsAsFactor}
\title{Get/Set Options for Encoding NASIS Domains as Factors}
\usage{
NASISDomainsAsFactor(x = NULL)
}
\arguments{
\item{x}{logical; default \code{FALSE}}
}
\value{
logical, result of \code{getOption("soilDB.NASIS.DomainsAsFactor")}
}
\description{
Set package option \code{soilDB.NASIS.DomainsAsFactor} for returning coded NASIS domains as factors.
}
\examples{
\dontrun{
NASISDomansAsFactor(TRUE)
}
}
