% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseWebReport.R
\name{parseWebReport}
\alias{parseWebReport}
\title{Parse contents of a web report, based on supplied arguments.}
\usage{
parseWebReport(url, args, index = 1)
}
\arguments{
\item{url}{Base URL to a LIMS/NASIS web report.}

\item{args}{List of named arguments to send to report, see details.}

\item{index}{Integer index specifying the table to return, or, NULL for a
list of tables}
}
\value{
A \code{data.frame} object in the case of a single integer \code{index}, otherwise a \code{list}
}
\description{
Parse contents of a web report, based on supplied arguments.
}
\details{
Report argument names can be inferred by inspection of the HTML source
associated with any given web report.
}
\note{
Most web reports are for internal use only.
}
\author{
D.E. Beaudette and S.M. Roecker
}
\keyword{IO}
