% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaExtent.R
\name{taxaExtent}
\alias{taxaExtent}
\title{Retrieve Soil Taxonomy Membership Grids}
\usage{
taxaExtent(
  x,
  level = c("order", "suborder", "greatgroup", "subgroup"),
  formativeElement = FALSE,
  timeout = 60
)
}
\arguments{
\item{x}{single taxon label (e.g. \code{haploxeralfs}) or formative element (e.g. \code{pale}), case-insensitive}

\item{level}{the taxonomic level within the top 4 tiers of Soil Taxonomy, one of \code{c('order', 'suborder', 'greatgroup', 'subgroup')}}

\item{formativeElement}{logical, search using formative elements instead of taxon label}

\item{timeout}{time that we are willing to wait for a response, in seconds}
}
\value{
a \code{raster} object
}
\description{
This function downloads a generalized representation of the geographic extent of any single taxon from the top 4 tiers of Soil Taxonomy, or those taxa matching a given formative element used in greatgroup or subgroup taxa. Data are provided by SoilWeb, ultimately sourced from from the current SSURGO snapshot. Data are returned as \code{raster} objects representing area proportion falling within 800m cells. Currently area proportions are based on major components only. Data are only available in CONUS and returned using an Albers Equal Area / NAD83 coordinate reference system (EPSG 6350).
}
\details{
\subsection{Taxon Queries}{

Taxon labels can be conveniently extracted from the "ST_unique_list" sample data, provided by the \href{https://github.com/ncss-tech/SoilTaxonomy}{SoilTaxonomy package}.
}

\subsection{Formative Element Queries}{
\subsection{Greatgroup:}{
\itemize{
\item acro: extreme weathering
\item alb: presence of an albic horizon
\item anhy: very dry
\item anthra: presence of an anthropic epipedon
\item aqui: wetness
\item argi: presence of an argillic horizon
\item calci: presence of a calcic horizon
\item cryo: cryic STR
\item duri: presence of a duripan
\item dystro: low base saturation
\item endo: ground water table
\item epi: perched water table
\item eutro: high base saturation
\item ferr: presence of Fe
\item fibr: least decomposed stage
\item fluv: flood plain
\item fol: mass of leaves
\item fragi: presence of a fragipan
\item fragloss: presence of a fragipan and glossic horizon
\item frasi: not salty
\item fulvi: dark brown with organic carbon
\item glac: presence of ice lenses
\item glosso: presence of a glossic horizon
\item gypsi: presence of a gypsic horizon
\item hal: salty
\item hapl: minimum horizon development
\item hemi: intermediate decomposition
\item histo: organic soil material
\item humi: presence of organic carbon
\item hydro: presence of water
\item kandi: presence of a kandic horizon
\item kanhaplo: thin kandic horizon
\item luvi: illuvial organic material
\item melano: presence of a melanic epipedon
\item molli: presence of a mollic epipedon
\item natri: presence of a natric horizon
\item pale: excessive development
\item petro: petrocalcic horizon
\item plac: presence of a thin pan
\item plagg: presence of a plaggen epipedon
\item plinth: presence of plinthite
\item psammo: sandy texture
\item quartzi: high quartz content
\item rhodo: dark red colors
\item sali: presence of a salic horizon
\item sapr: most decomposed stage
\item sombri: presence of a sombric horizon
\item sphagno: presence of sphagnum moss
\item sulfo: presence of sulfides or their oxidation products
\item torri: torric/aridic SMR
\item udi: udic SMR
\item umbri: presence of an umbric epipedon
\item usti: ustic SMR
\item verm: wormy, or mixed by animals
\item vitri: presence of glass
\item xero: xeric SMR
}
}

\subsection{Subgroup:}{
\itemize{
\item abruptic: abrupt textural change
\item acric: low apparent CEC
\item aeric: more aeration than typic subgroup
\item albaquic: presence of albic minerals, wetter than typic subgroup
\item albic: presence of albic minerals
\item alfic: presence of an argillic or kandic horizon
\item alic: high extractable Al content
\item anionic: low CEC or positively charged
\item anthraquic: human controlled flooding as in paddy rice culture
\item anthropic: an anthropic epipedon
\item aquic: wetter than typic subgroup
\item arenic: 50-100cm sandy textured surface
\item argic: argillic horizon
\item aridic: more aridic than typic subgroup
\item calcic: presence of a calcic horizon
\item chromic: high chroma colors
\item cumulic: thickened epipedon
\item duric: presence of a duripan
\item durinodic: presence of durinodes
\item dystric: lower base saturation percentage
\item entic: minimal surface/subsurface development
\item eutric: higher base saturation percentage
\item fibric: >25cm of fibric material
\item fluvaquentic: wetter than typic subgroup, evidence of stratification
\item fragiaquic: presence of fragic properties, wetter than typic subgroup
\item fragic: presence of fragic properties
\item glacic: presence of ice lenses or wedges
\item glossaquic: interfingered horizon boundaries, wetter than typic subgroup
\item glossic: interfingered horizon boundaries
\item grossarenic: >100cm sandy textured surface
\item gypsic: presence of gypsic horizon
\item halic: salty
\item haplic: central theme of subgroup concept
\item hemic: >25cm of hemic organic material
\item humic: higher organic matter content
\item hydric: presence of water
\item kandic: low activity clay present
\item lamellic: presence of lamellae
\item leptic: thinner than typic subgroup
\item limnic: presence of a limnic layer
\item lithic: shallow lithic contact present
\item natric: presence of sodium
\item nitric: presence of nitrate salts
\item ombroaquic: surface wetness
\item oxyaquic: water saturated but not reduced
\item pachic: epipedon thicker than typic subgroup
\item petrocalcic: presence of a petrocalcic horizon
\item petroferric: presence of petroferric contact
\item petrogypsic: presence of a petrogypsic horizon
\item petronodic: presence of concretions and/or nodules
\item placic: presence of a placic horizon
\item plinthic: presence of plinthite
\item rhodic: darker red colors than typic subgroup
\item ruptic: intermittent horizon
\item salic: presence of a salic horizon
\item sapric: >25cm of sapric organic material
\item sodic: high exchangeable Na content
\item sombric: presence of a sombric horizon
\item sphagnic: sphagnum organic material
\item sulfic: presence of sulfides
\item terric: mineral substratum within 1 meter
\item thaptic: presence of a buried soil horizon
\item turbic: evidence of cryoturbation
\item udic: more humid than typic subgroup
\item umbric: presence of an umbric epipedon
\item ustic: more ustic than typic subgroup
\item vermic: animal mixed material
\item vitric: presence of glassy material
\item xanthic: more yellow than typic subgroup
\item xeric: more xeric than typic subgroup
}
}

}
}
\examples{
\donttest{

if(requireNamespace("curl") &
   curl::has_internet()) {
  
  library(raster)
  
  # try a couple of different examples
  
  # soil order
  taxa <- 'vertisols'
  x <- taxaExtent(taxa, level = 'order')
  a <- raster::aggregate(x, fact = 5)
  
  # suborder
  taxa <- 'ustalfs'
  x <- taxaExtent(taxa, level = 'suborder')
  a <- raster::aggregate(x, fact = 5)
  
  # greatgroup
  taxa <- 'haplohumults'
  x <- taxaExtent(taxa, level = 'greatgroup')
  a <- raster::aggregate(x, fact = 5)
  
  # subgroup
  taxa <- 'Typic Haploxerepts'
  x <- taxaExtent(taxa, level = 'subgroup')
  a <- raster::aggregate(x, fact = 5)
  
  # quick evaluation of the result
  if(requireNamespace("rasterVis") & requireNamespace('viridisLite')) {
    rasterVis::levelplot(a, 
      margin = FALSE, scales = list(draw = FALSE), 
      col.regions = viridisLite::viridis, 
      main = names(a)
    )
  }
  
  # slippy map
  if(requireNamespace("mapview")) {
    mapview::mapview(a, col.regions = viridisLite::viridis, na.color = NA, use.layer.names = TRUE)
  }
  
  
  
}

}

}
\author{
D.E. Beaudette and Andrew Brown
}
