% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSURGO.R
\name{createSSURGO}
\alias{createSSURGO}
\title{Create a SpatiaLite or SQLite database from one or more SSURGO Exports}
\usage{
createSSURGO(filename, exdir, pattern = NULL)
}
\arguments{
\item{filename}{}

\item{exdir}{}

\item{pattern}{Character. Optional regular expression to use to filter subdirectories of \code{exdir}. Default: \code{NULL} will search all subdirectories for SSURGO export files.#' @param include_spatial Logical. Include spatial data layers in database? Default: \code{TRUE}.}
}
\value{
Logical. \code{TRUE} if all required SSURGO tables are successfully written to \code{filename}.
}
\description{
Create a SpatiaLite or SQLite database from one or more SSURGO Exports
}
