% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQueryNASIS.R
\name{dbConnectNASIS}
\alias{dbConnectNASIS}
\alias{NASIS}
\title{Create local NASIS database connection}
\usage{
dbConnectNASIS(dsn = NULL)

NASIS(dsn = NULL)
}
\arguments{
\item{dsn}{Optional: path to SQLite database containing NASIS table
structure; Default: \code{NULL}}
}
\value{
A \code{DBIConnection} object, as returned by
\code{DBI::dbConnect()}. If \code{dsn} is a \code{DBIConnection}, the attribute \code{isUserDefined} of the result is set to \code{TRUE}. If the \code{DBIConnection} is created by the internal NASIS connection process, \code{isUserDefined} is set to \code{FALSE.}
}
\description{
Create a connection to a local NASIS database with \code{DBI}
}
