% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLDM.R
\name{fetchLDM}
\alias{fetchLDM}
\title{Query data from Kellogg Soil Survey Laboratory Data Mart via Soil Data Access}
\usage{
fetchLDM(x, what = "pedlabsampnum", chunk.size = 1000)
}
\arguments{
\item{x}{a vector of values to find in column specified by \code{what}}

\item{what}{a single column name from either the \code{lab_combine_nasis_ncss} or the \code{lab_area} tables}

\item{chunk.size}{number of pedons per chunk (for queries that may exceed \code{maxJsonLength})}
}
\value{
a \code{SoilProfileCollection}
}
\description{
LDM model diagram:
}
\details{
\itemize{
\item https://jneme910.github.io/Lab_Data_Mart_Documentation/Documents/SDA_KSSL_Data_model.html
}
}
\examples{

# fetch by ssa_key
fetchLDM(8297, what = "ssa_key")

# get pedons correlated as taxonomic subgroup "Typic Argialbolls"
fetchLDM("Typic Argialbolls", what = "corr_taxsubgrp")
 
# TODO: generalize area_code functionality to include parent areas

## fetch by area_code (must be a soil survey area code for now; the lowest-level area that is always (?) populated)
# fetchLDM("CA630", what = "area_code")

}
