% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchNASISLabData.R
\name{fetchNASISLabData}
\alias{fetchNASISLabData}
\title{Fetch lab data used site/horizon data from a PedonPC database.}
\usage{
fetchNASISLabData(SS = TRUE, dsn = NULL)
}
\arguments{
\item{SS}{fetch data from the currently loaded selected set in NASIS or from
the entire local database (default: \code{TRUE})#'}

\item{dsn}{Optional: path to local SQLite database containing NASIS
table structure; default: \code{NULL}}
}
\value{
a SoilProfileCollection object
}
\description{
Fetch KSSL laboratory pedon/horizon layer data from a local NASIS database,
return as a SoilProfileCollection object.
}
\details{
This function currently works only on Windows, and requires a 'nasis_local'
ODBC connection.
}
\seealso{
\code{\link{get_labpedon_data_from_NASIS_db}}
}
\author{
J.M. Skovlin and D.E. Beaudette
}
\keyword{manip}
