% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSDA_spatial.R
\name{fetchSDA_spatial}
\alias{fetchSDA_spatial}
\title{Get Spatial Data from Soil Data Access by \code{mukey}, \code{nationalmusym} or \code{areasymbol}}
\usage{
fetchSDA_spatial(
  x,
  by.col = "mukey",
  method = "feature",
  geom.src = "mupolygon",
  db = "SSURGO",
  add.fields = NULL,
  chunk.size = 10,
  verbose = TRUE,
  as_Spatial = getOption("soilDB.return_Spatial", default = FALSE)
)
}
\arguments{
\item{x}{A vector of map unit keys (\code{mukey}) or national map unit symbols (\code{nationalmusym}) for \code{mupolygon}, \code{muline} or \code{mupoint}; feature keys (\code{featkey}) for \code{featpoint} and \code{featline}; legend keys (\code{lkey}) or soil survey area symbols (\code{areasymbol}) for \code{sapolygon} geometry. If \code{geom.src="mlrapolygon"} then \code{x} refers to \code{MLRARSYM} (major land resource area symbols).}

\item{by.col}{Column name containing map unit identifier \code{"mukey"}, \code{"nationalmusym"}, or \code{"ecoclassid"} for \code{geom.src} \code{mupolygon} OR \code{"areasymbol"}, \code{"areaname"}, \code{"mlraoffice"}, \code{"mouagncyresp"} for \code{geom.src} \code{sapolygon}; default is determined by \code{isTRUE(is.numeric(x))} for \code{mukey}, \code{featkey} or \code{lkey}, using \code{nationalmusym} or \code{areasymbol} otherwise.}

\item{method}{geometry result type: \code{"feature"} returns polygons, \code{"bbox"} returns the bounding box of each polygon (via \code{STEnvelope()}), \code{"point"} returns a single point (via \code{STPointOnSurface()}) within each polygon, \code{"extent"} returns an aggregate bounding box (the extent of all polygons, \code{geometry::EnvelopeAggregate()}) ), \code{"convexhull"} (\code{geometry::ConvexHullAggregate()}) returns the aggregate convex hull around all polygons, \code{"union"} (\code{geometry::UnionAggregate()}) and \code{"collection"} (\code{geometry::CollectionAggregate()}) return a \code{MULTIPOLYGON} or a \code{GEOMETRYCOLLECTION}, respectively, for each \code{mukey}, \code{nationalmusym}, or \code{areasymbol }. In the case of the latter four aggregation methods, the groups for aggregation depend on \code{by.col} (default by \code{"mukey"}).}

\item{geom.src}{Either \code{mupolygon} (map unit polygons), \code{muline} (map unit lines), \code{mupoint} (map unit points), \code{featpoint} (feature points), \code{featline} (feature lines), \code{sapolygon} (soil survey area boundary polygons), or \code{mlrapolygon} (major land resource area boundary polygons)}

\item{db}{Default: \code{"SSURGO"}. When \code{geom.src} is \code{mupolygon}, use STATSGO polygon geometry instead of SSURGO by setting \code{db = "STATSGO"}}

\item{add.fields}{Column names from \code{mapunit} or \code{legend} table to add to result. Must specify parent table name as the prefix before column name e.g. \code{mapunit.muname}.}

\item{chunk.size}{Number of values of \code{x} to process per query. Necessary for large results. Default: \code{10}}

\item{verbose}{Print messages?}

\item{as_Spatial}{Return sp classes? e.g. \code{Spatial*DataFrame}. Default: \code{FALSE}.}
}
\value{
an \code{sf} data.frame corresponding to SDA spatial data for all symbols requested. If \code{as_Spatial=TRUE} returns a \code{Spatial*DataFrame} from the sp package via \code{sf::as_Spatial()} for backward compatibility. Default result contains geometry with attribute table containing unique feature ID, symbol and area symbol plus additional fields in result specified with \code{add.fields}.
}
\description{
This method facilitates queries to Soil Data Access (SDA) mapunit and survey area geometry. Queries are generated based on map unit key (\code{mukey}) and national map unit symbol (\code{nationalmusym}) for \code{mupolygon} (SSURGO) or \code{gsmmupolygon} (STATSGO) geometry OR legend key (\code{lkey}) and area symbols (\code{areasymbol}) for \code{sapolygon} (Soil Survey Area; SSA) geometry).

A Soil Data Access query returns geometry and key identifying information about the map unit or area of interest. Additional columns from the map unit or legend table can be included; see \code{add.fields} argument.
}
\details{
This function automatically "chunks" the input vector (using \code{makeChunks()}) of map unit identifiers to minimize the likelihood of exceeding the SDA data request size. The number of chunks varies with the \code{chunk.size} setting and the length of your input vector. If you are working with many map units and/or large extents, you may need to decrease this number in order to have more chunks.

Querying regions with complex mapping may require smaller \code{chunk.size}. Numerically adjacent IDs in the input vector may share common qualities (say, all from same soil survey area or region) which could cause specific chunks to perform "poorly" (slow or error) no matter what the chunk size is. Shuffling the order of the inputs using \code{sample()} may help to eliminate problems related to this, depending on how you obtained your set of MUKEY/nationalmusym to query. One could feasibly use \code{muacres} as a heuristic to adjust for total acreage within chunks.

Note that STATSGO data are fetched where \code{CLIPAREASYMBOL = 'US'} to avoid duplicating state and national subsets of the geometry.

A prototype interface, \code{geom.src="mlrapolygon"}, is provided for obtaining Major Land Resource Area (MLRA) polygon
boundaries. When using this geometry source \code{x} is a vector of \code{MLRARSYM} (MLRA Symbols). The geometry source is
the MLRA Geographic Database v5.2 (2022) which is not (yet) part of Soil Data Access. Instead of SDA, GDAL utilities
are used to read a zipped ESRI Shapefile from a remote URL: \url{https://www.nrcs.usda.gov/sites/default/files/2022-10/MLRA_52_2022.zip}.
Therefore, most additional \code{fetchSDA_spatial()} arguments are \emph{not} currently supported for the MLRA geometry source.
In the future a \code{mlrapolygon} table may be added to SDA (analogous to  \code{mupolygon} and \code{sapolygon}),
and the function will be updated accordingly at that time.
}
\examples{
\dontshow{if (curl::has_internet() && requireNamespace("httr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
 
    # get spatial data for a single mukey
    single.mukey <- try(fetchSDA_spatial(x = "2924882"))

    # demonstrate fetching full extent (multi-mukey) of national musym
    full.extent.nmusym <- try(fetchSDA_spatial(x = "2x8l5", by = "nmusym"))

    # compare extent of nmusym to single mukey within it
    if (!inherits(single.mukey, 'try-error') && 
        !inherits(full.extent.nmusym, 'try-error')) {
        
        if (requireNamespace("sf")) {
      
         plot(sf::st_geometry(full.extent.nmusym), col = "RED", border = 0)
         plot(sf::st_geometry(single.mukey), add = TRUE, col = "BLUE", border = 0)
       
        }
        
    }

    # demo adding a field (`muname`) to attribute table of result
    head(try(fetchSDA_spatial(x = "2x8l5", by="nmusym", add.fields="muname")))
}
\dontshow{\}) # examplesIf}
}
\author{
Andrew G. Brown, Dylan E. Beaudette
}
