% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSRI.R
\name{fetchSRI}
\alias{fetchSRI}
\title{Fetch Soil Inventory Resource (SRI) for USFS Region 6}
\usage{
fetchSRI(gdb, ...)
}
\arguments{
\item{gdb}{A \code{character} vector of the gdb(s), e.g. \code{'Deschutes'}.}

\item{...}{Arguments to pass to \link{get_SRI}.}
}
\value{
A list.
}
\description{
Fetch Soil Inventory Resource (SRI) for USFS Region 6
}
\examples{
\dontshow{if (curl::has_internet() && requireNamespace("sf")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{

# fetch Willamette and Winema SRI

sri <- fetchSRI(gdb = c('will', 'win'), quiet = TRUE)

}
\dontshow{\}) # examplesIf}
}
