% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors_from_NASIS_db.R
\name{get_colors_from_NASIS_db}
\alias{get_colors_from_NASIS_db}
\title{Extract Soil Color Data from a local NASIS Database}
\usage{
get_colors_from_NASIS_db(SS = TRUE, dsn = NULL)
}
\arguments{
\item{SS}{fetch data from Selected Set in NASIS or from the entire local
database (default: \code{TRUE})}

\item{dsn}{Optional: path to local SQLite database containing NASIS
table structure; default: \code{NULL}}
}
\value{
A data.frame with the results.
}
\description{
Get, format, mix, and return color data from a NASIS database.
}
\details{
This function currently works only on Windows.
}
\seealso{
\code{\link{simplifyColorData}},
\code{\link{get_hz_data_from_NASIS_db}},
\code{\link{get_site_data_from_NASIS_db}}
}
\author{
Jay M. Skovlin and Dylan E. Beaudette
}
\keyword{manip}
