% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchEDIT_tools.R
\name{get_EDIT_ecoclass_by_geoUnit}
\alias{get_EDIT_ecoclass_by_geoUnit}
\title{Get data.frame of all ecoclass for multiple EDIT geoUnit}
\usage{
get_EDIT_ecoclass_by_geoUnit(geoUnit)
}
\arguments{
\item{geoUnit}{A character vector of `geoUnit` codes e.g. `c("018X","022A")` for MLRAs 18 and 22A.}
}
\value{
A `data.frame` containing: `geoUnit`, `id`, `legacyId`, `name`
}
\description{
Supply a vector of target `geoUnit`. Data are accessed via Ecological Dynamics Interpretive Tool (EDIT) web services: https://edit.jornada.nmsu.edu/resources/esd
}
\examples{

get_EDIT_ecoclass_by_geoUnit(c("018X","022A"))

}
