\name{get_extended_data_from_NASIS_db}
\alias{get_extended_data_from_NASIS_db}

\title{Extract accessory tables and summaries from a local NASIS Database}
\description{Extract accessory tables and summaries from a local NASIS Database.}
\usage{
get_extended_data_from_NASIS_db()
}

\details{This function currently works only on Windows.}
\value{A list with the results.}
\author{Jay M. Skovlin and Dylan E. Beaudette}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{get_hz_data_from_NASIS_db}}, \code{\link{get_site_data_from_NASIS_db}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}
