\name{get_labpedon_data_from_NASIS_db}
\alias{get_labpedon_data_from_NASIS_db}

\title{Extract lab pedon data from a local NASIS Database}
\description{Get lab pedon-level data from a local NASIS database.}
\usage{get_labpedon_data_from_NASIS_db()}

\value{A dataframe.}
\author{Jay M. Skovlin and Dylan E. Beaudette}
\details{}
\note{This function currently works only on Windows.}

\seealso{
\code{\link{get_lablayer_data_from_NASIS_db}}, 
}

\examples{

\dontrun{


}}

\keyword{manip}

