% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapunit_from_NASIS.R,
%   R/get_projectmapunit_from_NASIS.R
\name{get_mapunit_from_NASIS}
\alias{get_mapunit_from_NASIS}
\alias{get_legend_from_NASIS}
\alias{get_lmuaoverlap_from_NASIS}
\alias{get_projectmapunit_from_NASIS}
\title{Get Legend, Mapunit and Legend Mapunit Area Overlap Tables}
\usage{
get_mapunit_from_NASIS(
  SS = TRUE,
  droplevels = TRUE,
  stringsAsFactors = NULL,
  dsn = NULL
)

get_legend_from_NASIS(
  SS = TRUE,
  droplevels = TRUE,
  stringsAsFactors = NULL,
  dsn = NULL
)

get_lmuaoverlap_from_NASIS(
  SS = TRUE,
  droplevels = TRUE,
  stringsAsFactors = NULL,
  dsn = NULL
)

get_projectmapunit_from_NASIS(SS = TRUE, stringsAsFactors = NULL, dsn = NULL)
}
\arguments{
\item{SS}{fetch data from the currently loaded selected set in NASIS or from the entire local database (default: \code{TRUE})}

\item{droplevels}{Drop unused levels from \code{farmlndcl} and other factor levels from NASIS domains?}

\item{stringsAsFactors}{deprecated}

\item{dsn}{Optional: path to local SQLite database containing NASIS
table structure; default: \code{NULL}}
}
\description{
Get Legend, Mapunit and Legend Mapunit Area Overlap Tables
}
