% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_hydric.R
\name{get_SDA_hydric}
\alias{get_SDA_hydric}
\title{Get map unit hydric soils information from Soil Data Access}
\usage{
get_SDA_hydric(areasymbols = NULL, mukeys = NULL)
}
\arguments{
\item{areasymbols}{vector of soil survey area symbols}

\item{mukeys}{vector of map unit keys}
}
\value{
a data.frame
}
\description{
Get map unit hydric soils information from Soil Data Access
}
\author{
Jason Nemecek, Chad Ferguson, Andrew Brown
}
