% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_interpretations.R
\name{get_SDA_interpretation}
\alias{get_SDA_interpretation}
\title{Get map unit interpretations from Soil Data Access by rule name}
\usage{
get_SDA_interpretation(rulename, method, areasymbols = NULL, mukeys = NULL)
}
\arguments{
\item{rulename}{rule name of interpretation}

\item{method}{aggregation method. One of: "Dominant Component", "Dominant Condition", "Weighted Average"}

\item{areasymbols}{vector of soil survey area symbols}

\item{mukeys}{vector of map unit keys}
}
\value{
a data.frame
}
\description{
Get map unit interpretations from Soil Data Access by rule name
}
\author{
Jason Nemecek, Chad Ferguson, Andrew Brown
}
