% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilDataViewer.R
\name{get_SDV_legend_elements}
\alias{get_SDV_legend_elements}
\title{Get Soil Data Viewer Attribute Information}
\usage{
get_SDV_legend_elements(
  WHERE,
  alpha = 255,
  notratedcolor = rgb(1, 1, 1, 0),
  simplify = TRUE
)
}
\arguments{
\item{WHERE}{WHERE clause for query of Soil Data Access \code{sdvattribute} table}

\item{alpha}{transparency value applied in calculation of hexadecimal color. Default: \code{255} (opaque).}

\item{notratedcolor}{Used to add 'Not rated' color entries where applicable. Default: \code{"#FFFFFF00"} (transparent white).}

\item{simplify}{Return a data.frame when \code{WHERE} is length 1? Return a list with 1 element per legend when \code{WHERE} is length > \code{1}? Default: \code{TRUE}}
}
\value{
A list with a data.frame element for each element of \code{where} containing \code{"attributekey"}, \code{"attributename"}, \code{"attributetype"}, \code{"attributetablename"}, \code{"attributecolumnname"}, \code{"attributedescription"}, \code{"nasisrulename"}, \code{"label"}, \code{"order"}, \code{"value"}, \code{"lower_value"}, \code{"upper_value"},\code{"red"}, \code{"green"}, \code{"blue"} and \code{"hex"} columns.
}
\description{
Get Soil Data Viewer Attribute Information
}
