\name{get_site_data_from_NASIS_db}
\alias{get_site_data_from_NASIS_db}

\title{Extract Site Data from a local NASIS Database}
\description{Get site-level data from a local NASIS database.}
\usage{
get_site_data_from_NASIS_db(dsn)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dsn}{The name of the local NASIS database connection (see details).}
}
\details{}
\value{A dataframe.}
\references{
%% ~put references to the literature/web site here ~
}
\author{Jay M. Skovlin and Dylan E. Beaudette}
\note{This function currently works only on Windows.}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{get_hz_data_from_NASIS_db}}, \code{\link{get_veg_from_AK_Site}}, 
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}

