\name{get_site_data_from_pedon_db}
\alias{get_site_data_from_pedon_db}

\title{Extract Site Data from a PedonPC Database}
\description{Get site-level data from a PedonPC database.}
\usage{
get_site_data_from_pedon_db(dsn)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dsn}{The path to a 'pedon.mdb' database.}
}
\details{}
\value{A dataframe.}
\references{
%% ~put references to the literature/web site here ~
}
\author{Dylan E. Beaudette and Jay Skovlin}
\note{This function currently works only on Windows.}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{get_hz_data_from_pedon_db}}, \code{\link{get_veg_from_AK_Site}}, 
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}

