% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{soilDB.env}
\alias{soilDB.env}
\alias{get_soilDB_env}
\title{Get the soilDB environment used for storing error messages and quality control output}
\format{
An object of class \code{environment} of length 0.
}
\usage{
soilDB.env

get_soilDB_env()
}
\value{
a \code{environment} object
}
\description{
The soilDB package uses an environment to store variables that are created as side effects of various data access and processing routines.
\code{get_soilDB_env()} provides a method to access this environment from the global (user) environment.
}
\examples{
get_soilDB_env()
}
\keyword{datasets}
