% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soilseries_from_NASIS.R
\name{get_soilseries_from_NASIS}
\alias{get_soilseries_from_NASIS}
\alias{get_soilseries_from_NASISWebReport}
\alias{get_competing_soilseries_from_NASIS}
\title{Get records from the Series Classification (SC) database}
\usage{
get_soilseries_from_NASIS(dsn = NULL, delimiter = " over ", SS = FALSE)

get_soilseries_from_NASISWebReport(soils)

get_competing_soilseries_from_NASIS(
  x,
  what = "taxclname",
  dsn = NULL,
  SS = FALSE
)
}
\arguments{
\item{dsn}{Optional: path or \emph{DBIConnection} to \link[=NASISLocalDatabase]{local database containing NASIS table structure}; default: \code{NULL}}

\item{delimiter}{\emph{character}. Used to collapse \code{taxminalogy} records where multiple values are used to describe strongly contrasting control sections. Default \code{" over "} creates combination mineralogy classes as they would be used in the family name.}

\item{SS}{\emph{logical}. Fetch data from the currently loaded selected set in NASIS or from the entire local database (default: \code{FALSE}; this is to allow for queries against the full Series Classification database as default)}

\item{soils}{A vector of soil series names}

\item{x}{Taxonomic Class Name (or other field specified by \code{what}) to match, use \verb{\%} for wildcard}

\item{what}{Column name to match \code{x} against, default: \code{'taxclname'}}
}
\value{
A \code{data.frame}
}
\description{
These functions return records from the Series Classification (SC) database, either
from the local NASIS database (all series) or via web report (named series
only).

\verb{get_competing_soilseries_from_NASIS():} Get Soil Series from NASIS Matching Taxonomic Class Name
}
\author{
Stephen Roecker
}
\keyword{manip}
