% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soilseries_from_NASIS.R
\name{get_soilseries_from_NASIS}
\alias{get_soilseries_from_NASIS}
\alias{get_soilseries_from_NASISWebReport}
\title{Get records from the Soil Classification (SC) database}
\usage{
get_soilseries_from_NASIS(
  stringsAsFactors = default.stringsAsFactors(),
  dsn = NULL
)
}
\arguments{
\item{stringsAsFactors}{logical: should character vectors be converted to
factors? This argument is passed to the \code{uncode()} function. It does not
convert those vectors that have set outside of \code{uncode()} (i.e. hard coded).}

\item{dsn}{Optional: path to local SQLite database containing NASIS
table structure; default: \code{NULL}}
}
\value{
A \code{data.frame}
}
\description{
These functions return records from the Soil Classification database, either
from the local NASIS database (all series) or via web report (named series
only).
}
\author{
Stephen Roecker
}
\keyword{manip}
