% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSRI.R
\name{get_SRI_layers}
\alias{get_SRI_layers}
\title{Get SRI Layers}
\usage{
get_SRI_layers(gdb)
}
\arguments{
\item{gdb}{A \code{character} of the GDB, e.g. \code{'Deschutes'}.}
}
\value{
A list of metadata about the GDB
}
\description{
Get SRI Layers
}
\note{
Refer to \code{\link{get_SRI}} for information on File Geodatabase (GDB) availability.
}
\examples{
\dontrun{
sri_layers <- get_SRI_layers('Willamette')
}

}
\author{
Josh Erickson
}
