% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHzErrorsNASIS.R
\name{getHzErrorsNASIS}
\alias{getHzErrorsNASIS}
\title{Get Logic Errors in NASIS/PedonPC Pedon Horizon}
\usage{
getHzErrorsNASIS(strict = TRUE, SS = TRUE, dsn = NULL)
}
\arguments{
\item{strict}{how strict should horizon boundaries be checked for
consistency: TRUE=more | FALSE=less}

\item{SS}{fetch data from the currently loaded selected set in NASIS or from
the entire local database (default: TRUE)}

\item{dsn}{Optional: path to local SQLite database containing NASIS
table structure; default: NULL}
}
\value{
A data.frame containing problematic records with columns:
'peiid','upedonid','hzdept','hzdepb','hzname'
}
\description{
Get Logic Errors in NASIS/PedonPC Pedon Horizon
}
