% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISSR800.R
\name{ISSR800.wcs}
\alias{ISSR800.wcs}
\title{ISSR-800 Web Coverage Service Interface}
\usage{
ISSR800.wcs(var, aoi, res = 800, crs = "EPSG:6350")
}
\arguments{
\item{var}{variable label}

\item{aoi}{area of interest as WGS84 coordinates c(-121, 37, -120, 38) (xmin, ymin, xmax, ymax)}

\item{res}{grid cell resolution (units specific to \code{crs}), typically 800 (meters), the native resolution of ISSR-800}

\item{crs}{coordinate reference specification in the form `EPSG:6350`, must contain a valid EPSG code}
}
\value{
\code{raster} object
}
\description{
ISSR-800 Web Coverage Service Interface
}
\note{
This is an experimental interface that can change at any time.
}
