\name{loafercreek}
\alias{loafercreek}
\docType{data}
\title{LoaferCreek Soils}
\description{A SoilProfileCollection Object of LoaferCreek Soils from the CA630 Soil Survey Area.}
\usage{data(loafercreek)}
\format{\preformatted{Formal class 'SoilProfileCollection' [package "aqp"] with 7 slots
  ..@ idcol     : chr "pedon_id"
  ..@ depthcols : chr [1:2] "hzdept" "hzdepb"
  ..@ metadata  :'data.frame':	1 obs. of  1 variable:
  .. ..$ depth_units: chr "cm"
  ..@ horizons  :'data.frame':	281 obs. of  20 variables:
  .. ..$ phiid          : int [1:281] 2484 2485 2486 2487 2488 2489 2490 2510 2511 2512 ...
  .. ..$ pedon_id       : chr [1:281] "09BAH006" "09BAH006" "09BAH006" "09BAH006" ...
  .. ..$ hzname         : chr [1:281] "Oi" "A" "Bt1" "Bt2" ...
  .. ..$ hzdept         : int [1:281] 0 1 6 18 29 40 59 0 1 4 ...
  .. ..$ hzdepb         : int [1:281] 1 6 18 29 40 59 64 1 4 12 ...
  .. ..$ clay           : num [1:281] NA 18 20 24 30 42 NA NA 16 18 ...
  .. ..$ silt           : num [1:281] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ sand           : num [1:281] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ phfield        : num [1:281] NA 6.5 6.5 6.5 6.5 ...
  .. ..$ effervescence  : chr [1:281] NA NA NA NA ...
  .. ..$ labsampnum     : logi [1:281] NA NA NA NA NA NA ...
  .. ..$ total_frags_pct: num [1:281] 0 5 5 10 16 61 0 0 10 10 ...
  .. ..$ texture_class  : chr [1:281] "spm" "l" "l" "l" ...
  .. ..$ d_r            : num [1:281] NA 0.599 0.617 0.637 0.637 ...
  .. ..$ d_g            : num [1:281] NA 0.457 0.448 0.441 0.441 ...
  .. ..$ d_b            : num [1:281] NA 0.266 0.291 0.248 0.248 ...
  .. ..$ m_r            : num [1:281] NA 0.382 0.492 0.51 0.51 ...
  .. ..$ m_g            : num [1:281] NA 0.26 0.352 0.346 0.346 ...
  .. ..$ m_b            : num [1:281] NA 0.0932 0.2313 0.1953 0.1953 ...
  .. ..$ soil_color     : chr [1:281] NA "#614218" "#7E5A3B" "#825832" ...
  ..@ site      :'data.frame':	48 obs. of  24 variables:
  .. ..$ pedon_id       : chr [1:48] "09BAH006" "09BAH008" "09CKS006" "09CKS010" ...
  .. ..$ peiid          : int [1:48] 785 787 767 771 809 812 898 1155 1164 1192 ...
  .. ..$ site_id        : chr [1:48] "09CA630BAH006" "09CA630BAH008" "09CA630CKS006" "09CA630CKS010" ...
  .. ..$ siteiid        : int [1:48] 760 762 742 746 784 787 867 1122 1131 1159 ...
  .. ..$ sampled_as     : chr [1:48] "Loafercreek" "Loafercreek" "Loafercreek" "Loafercreek" ...
  .. ..$ correlated_as  : chr [1:48] "Loafercreek" "Loafercreek" "LOAFERCREEK" "LOAFERCREEK" ...
  .. ..$ hillslope_pos  : chr [1:48] "backslope" "footslope" "backslope" "backslope" ...
  .. ..$ loafercreek              : num [1:48] -121 -121 -121 -121 -121 ...
  .. ..$ y              : num [1:48] 38 38 37.9 37.9 37.9 ...
  .. ..$ datum          : chr [1:48] "NAD83" "NAD83" "NAD83" "NAD83" ...
  .. ..$ elev           : num [1:48] 450 382 235 207 338 445 233 362 481 413 ...
  .. ..$ slope          : num [1:48] 32 3 29 4 36 50 18 32 11 8 ...
  .. ..$ aspect         : int [1:48] 248 102 235 115 185 185 335 52 285 40 ...
  .. ..$ plantassocnm   : chr [1:48] "Grassland" "Blue Oak Woodland/Grassland" "Blue Oak Woodlland" "Live Oak Woodland/Grassland" ...
  .. ..$ bedrckdepth    : int [1:48] 59 68 86 96 72 93 65 70 60 40 ...
  .. ..$ bedrock_kind   : chr [1:48] "Metavolcanics" "Metavolcanics" "Greenstone" "Greenstone" ...
  .. ..$ describer      : chr [1:48] "Bev Harben" "Bev Harben" "Chris Savastio" "Chris Savastio" ...
  .. ..$ psctopdepth    : int [1:48] 6 12 3 2 12 8 5 25 8 14 ...
  .. ..$ pscbotdepth    : int [1:48] 56 62 53 52 57 58 55 70 60 40 ...
  .. ..$ part_size_class: chr [1:48] "fine-loamy" "fine-loamy" "fine-loamy" "fine-loamy" ...
  .. ..$ tax_subgroup   : chr [1:48] "Ultic Haploxeralfs" "Ultic Haploxeralfs" "Ultic Haploxeralfs" "Ultic Haploxeralfs" ...
  .. ..$ obs_date       : POSIXct[1:48], format: "2009-06-02 12:42:00" "2009-06-03 10:27:00" "2009-04-14 13:57:00" ...
  .. ..$ pedon_purpose  : chr [1:48] "full pedon description" "full pedon description" NA "forestry data site" ...
  .. ..$ pedlabsampnum  : chr [1:48] NA NA NA NA ...
  ..@ sp        :Formal class 'SpatialPoints' [package "sp"] with 3 slots
  .. .. ..@ coords     : num [1, 1] 0
  .. .. ..@ bbox       : logi [1, 1] NA
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slots
  .. .. .. .. ..@ projargs: chr NA
  ..@ diagnostic:'data.frame':	0 obs. of  0 variables
}}
\details{}
\source{}
\references{}
\examples{
data(loafercreek)

##
## aggregate major horizon types over 1cm intervals
##

# categorize major horizon types
hz.tab <- rev(sort(table(loafercreek$hzname)))
hz.tab[hz.tab > 5]

# add generalized hz name
loafercreek$hz <- rep('other', times=nrow(horizons(loafercreek)))

# generalize horizons
loafercreek$hz[grep('O', loafercreek@horizons$hzname)] <- 'O'
loafercreek$hz[grep('A', loafercreek@horizons$hzname)] <- 'A'
loafercreek$hz[grep('Bt', loafercreek@horizons$hzname)] <- 'Bt'
loafercreek$hz[grep('Cr', loafercreek@horizons$hzname)] <- 'Cr'
loafercreek$hz[grep('R', loafercreek@horizons$hzname)] <- 'R'

# convert generalized hz to factor
loafercreek$hz <- factor(loafercreek$hz)
loafercreek.hz.agg <- slab(loafercreek, fm= ~ hz)

# wide->long format
loafercreek.hz.agg.long <- melt(loafercreek.hz.agg, id.var=c('top','bottom', 'contributing_fraction', 'all.profiles', 'variable'),  variable_name='horizon')

# plot horizon type proportions
xyplot(top ~ value, groups=horizon, data=loafercreek.hz.agg.long, ylim=c(150, -5), type=c('S','g'), horizontal=TRUE, subset=value > 0 & horizon != 'other', asp=2, ylab='', xlab='')
}
\keyword{datasets}
