% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mukey-WCS.R
\name{mukey.wcs}
\alias{mukey.wcs}
\title{gNATSGO / gSSURGO Web Coverage Service Interface}
\usage{
mukey.wcs(var = "gnatsgo", aoi, res = 30, crs = "EPSG:6350")
}
\arguments{
\item{var}{variable label}

\item{aoi}{area of interest as WGS84 coordinates c(-121, 37, -120, 38) (xmin, ymin, xmax, ymax)}

\item{res}{grid cell resolution (units specific to \code{crs}), typically 30 (meters), the native resolution of gNATSGO and gSSURGO}

\item{crs}{coordinate reference specification in the form `EPSG:6350`, must contain a valid EPSG code}
}
\value{
\code{raster} object containing indexed map unit keys and associated raster attribute table
}
\description{
gNATSGO / gSSURGO Web Coverage Service Interface
}
\note{
This is an experimental interface that can change at any time.
}
