% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSDquery.R
\name{OSDquery}
\alias{OSDquery}
\title{Search full text of Official Series Description on SoilWeb}
\usage{
OSDquery(
  everything = NULL,
  mlra = "",
  taxonomic_class = "",
  typical_pedon = "",
  brief_narrative = "",
  ric = "",
  use_and_veg = "",
  competing_series = "",
  geog_location = "",
  geog_assoc_soils = "",
  remarks = ""
)
}
\arguments{
\item{everything}{search entire OSD text (default is NULL), \code{mlra} may also be specified, all other arguments are ignored}

\item{mlra}{a comma-delimited string of MLRA to search ('17,18,22A'), see Details}

\item{taxonomic_class}{search family level classification, see Details}

\item{typical_pedon}{search typical pedon section}

\item{brief_narrative}{search brief narrative}

\item{ric}{search range in characteristics section}

\item{use_and_veg}{search use and vegetation section}

\item{competing_series}{search competing series section}

\item{geog_location}{search geographic setting section}

\item{geog_assoc_soils}{search geographically associated soils section}

\item{remarks}{search remarks section (typically contains diagnostic horizons / features)}
}
\value{
a \code{data.frame} object containing soil series names that match patterns supplied as arguments.
}
\description{
This is the R interface to \href{https://casoilresource.lawr.ucdavis.edu/osd-search/}{OSD search by Section} and \href{https://casoilresource.lawr.ucdavis.edu/osd-search/search-entire-osd.php}{OSD Search} APIs provided by SoilWeb.

OSD records are searched with the \href{https://www.postgresql.org/docs/current/textsearch.html}{PostgreSQL fulltext indexing} and query system (\href{https://www.postgresql.org/docs/current/textsearch-controls.html}{syntax details}). Each search field (except for the "brief narrative" and MLRA) corresponds with a section header in an OSD. The results may not include every OSD due to formatting errors and typos. Results are scored based on the number of times search terms match words in associated sections.
}
\details{
Queries including the \code{taxonomic_class} argument make use of the Soil Classification database, not fulltext search of OSD records. Queries including the \code{mlra} argument make use of a SoilWeb data source based on spatial intersection (SSURGO x MLRA polygons), updated quarterly. MLRA queries are only possible for those soil series used in the current SSURGO snapshot.

The \code{mlra} argument must be combined with another argument in order to become active. For example, search for series with "5GY" hues in the "typical pedon" section, but limit to just MLRA 18: \code{OSDquery(mlra = '18', typical_pedon = '5GY')}.
\subsection{Syntax Notes:}{

The PostgreSQL fulltext query syntax is complex, but many common text search concepts are familiar:
\itemize{
\item logical AND: \code{&}
\item logical OR: \code{|}
\item wildcard, e.g. rhy-something: \verb{rhy:*}
\item search terms with spaces need doubled single quotes: \verb{''san joaquin''}
\item combine search terms into a single expression: \verb{(grano:* | granite)}
}
}

\subsection{Examples}{

Strategies for searching entire OSD records:
\itemize{
\item \verb{iowa & smectitic & verti:* & Cg & ! saturated}
\item \verb{iowa & smectitic & verti:* & Cg & terrace}
\item \verb{(sulfi:* | sulfa:*) & aq:*}
\item \code{Coarse-loamy & mixed & active & thermic & Mollic & Haploxeralfs}
\item \verb{sierra & nevada & (meta:* | metamorphic) & xer:* & thermic & lithic}
\item \verb{sierra & nevada & foothill & (grano:* | granite) & thermic}
\item \verb{rhyo:* & tuff:* & California & thermic}
\item \code{paralithic & thermic & !mesic & mollic & epipedon}
\item \verb{(gypsum | gyp:*) (MLRA: 15,17)}
\item \code{flood & plains & toe & slope}
}

Strategies for search OSD fields:
\itemize{
\item \code{taxonomic_class = 'duri:* & thermic'}: family level classification contains "duri-" prefix and "thermic"
\item \code{typical_pedon = 'cobbly & ashy & silt & loam'}: "cobbly, ashy, silt, loam", any horizon
\item \code{geog_location = 'strath & terrace'}: "strath" and "terrace" in geographic setting narrative
}

Related documentation can be found in the following tutorials
\itemize{
\item \href{http://ncss-tech.github.io/AQP/soilDB/soil-series-query-functions.html}{overview of all soil series query functions}
\item \href{https://ncss-tech.github.io/AQP/soilDB/competing-series.html}{competing soil series}
\item \href{https://ncss-tech.github.io/AQP/soilDB/siblings.html}{siblings}
}
}
}
\note{
SoilWeb maintains a snapshot of the Official Series Description data, updated quarterly.
}
\examples{
\dontshow{if (curl::has_internet() && requireNamespace("httr") && requireNamespace("jsonlite") && require(aqp)) withAutoprint(\{ # examplesIf}

\donttest{
  # find all series that list Pardee as a geographically associated soil.
  s <- OSDquery(geog_assoc_soils = 'pardee')

  # get data for these series
  x <- fetchOSD(s$series, extended = TRUE, colorState = 'dry')

  # simple figure
  par(mar = c(0,0,1,1))
  aqp::plotSPC(x$SPC)
}
\dontshow{\}) # examplesIf}
}
\references{
USDA-NRCS OSD search tools: \url{https://soilseries.sc.egov.usda.gov/}
}
\seealso{
\code{\link[=fetchOSD]{fetchOSD()}}, \code{\link[=siblings]{siblings()}}
}
\author{
D.E. Beaudette
}
\keyword{manip}
