% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSETTA.R
\name{ROSETTA}
\alias{ROSETTA}
\title{ROSETTA Model Versions 1 and 3 API}
\usage{
ROSETTA(x, v = c("1", "3"), conf = NULL)
}
\arguments{
\item{x}{a \code{data.frame} or \code{matrix} of soil properties, may not contain NA, see details}

\item{v}{single character of '1' or '3', this is the model version number}

\item{conf}{configuration passed to code{httr::POST()}.}
}
\value{
a \code{data.frame} object with estimated water retention curve parameters and saturated hydraulic conductivity:

\describe{

 \item{theta_r: }{residual volumetric water content (cm^3/cm^3)}
 \item{theta_s: }{saturated volumetric water content (cm^3/cm^3)}
 \item{alpha:}{related to the inverse of the air entry suction, log10-tranformed values with units of cm}
 \item{npar: }{index of pore size distribution, log10-tranformed values with units of 1/cm}
 \item{ksat: }{saturated hydraulic conductivity, log10-transformed values with units of cm/day}
 
}
}
\description{
A simple interface to the ROSETTA model for predicting hydraulic parameters from soil properties. Access is via API and still a work in progress.
}
\details{
...
}
