% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDB-package.R
\docType{data}
\name{SCAN_SNOTEL_metadata}
\alias{SCAN_SNOTEL_metadata}
\alias{state_FIPS_codes}
\title{Get SCAN and SNOTEL Station Metadata}
\format{
A data frame with 1092 observations on the following 12 variables.
\describe{ \item{list("Name")}{station name} \item{list("Site")}{station ID}
\item{list("State")}{state} \item{list("Network")}{sensor network: SCAN /
SNOTEL} \item{list("County")}{county} \item{list("Elevation_ft")}{station
elevation in feet} \item{list("Latitude")}{latitude of station}
\item{list("Longitude")}{longitude of station} \item{list("HUC")}{associated
watershed} \item{list("climstanm")}{climate station name (TODO: remove this
column)} \item{list("upedonid")}{associated user pedon ID}
\item{list("pedlabsampnum")}{associated lab sample ID} }
}
\description{
These data have been compiled from several sources and represent a
progressive effort to organize SCAN/SNOTEL station metadata. Therefore, some
records may be missing or incorrect.
}
\keyword{datasets}
