% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDB-package.R
\docType{package}
\name{soilDB-package}
\alias{soilDB-package}
\alias{soilDB.env}
\alias{soilDB}
\title{Soil Database Interface}
\description{
This package provides methods for extracting soils information from local
PedonPC and AK Site databases (MS Access format), local NASIS databases (MS
SQL Server), and the SDA web service. Currently USDA-NCSS data sources are
supported, however, there are plans to develop interfaces to outside systems
such as the Global Soil Mapping project.
}
\seealso{
\code{\link{fetchPedonPC}, \link{fetchNASIS}, \link{SDA_query}, \link{loafercreek}}
}
\author{
J.M. Skovlin, D.E. Beaudette, S.M Roecker, A.G. Brown
}
\keyword{package}
