% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDB-package.R
\docType{package}
\name{soilDB-package}
\alias{soilDB-package}
\alias{soilDB.env}
\alias{soilDB}
\title{Soil Database Interface}
\description{
A collection of functions for reading data from USDA-NCSS soil databases.
}
\details{
This package provides methods for extracting soils information from local
PedonPC and AK Site databases (MS Access format), local NASIS databases (MS
SQL Server), Soil Data Access and various other soil-related web services.
}
\seealso{
\code{\link{fetchPedonPC}, \link{fetchNASIS}, \link{SDA_query}, \link{loafercreek}}
}
\author{
J.M. Skovlin, D.E. Beaudette, S.M Roecker, A.G. Brown
}
\keyword{package}
