\name{soilDB-package}
\alias{soilDB-package}
\alias{soilDB}
\docType{package}
\title{Soil Database Interface}
\description{A collection of functions for reading data from USDA-NCSS soil databases.}
\details{Overview}
\author{
D.E. Beaudette and J.M. Skovlin
Maintainer: D.E. Beaudette <dylan.beaudette@ca.usda.gov>
}
\references{}
\keyword{package}
\seealso{}
\examples{}
