\name{soilDB-package}
\alias{soilDB-package}
\alias{soilDB}
\docType{package}
\title{Soil Database Interface}
\description{A collection of functions for reading data from USDA-NCSS soil databases.}
\details{It can be difficult to locate all of the dependencies required for sending/processing SOAP requests. install.packages("XMLSchema", repos = "http://www.omegahat.org/R", type = "source")}
\author{
D.E. Beaudette and J.M. Skovlin
Maintainer: D.E. Beaudette <dylan.beaudette@ca.usda.gov>
}
\references{}
\keyword{package}
\seealso{}
\examples{}
