% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaExtent.R
\name{taxaExtent}
\alias{taxaExtent}
\title{Retrieve Soil Taxonomy Membership Grids}
\usage{
taxaExtent(
  x,
  level = c("order", "suborder", "greatgroup", "subgroup"),
  timeout = 60
)
}
\arguments{
\item{x}{single taxa name, case-insensitive}

\item{level}{the taxonomic level within the top 4 tiers of Soil Taxonomy, one of \code{c('order', 'suborder', 'greatgroup', 'subgroup')}}

\item{timeout}{time that we are willing to wait for a response, in seconds}
}
\value{
a \code{raster} object
}
\description{
This function downloads a generalized representation of the geographic extent of any single taxa from the top 4 tiers of Soil Taxonomy. Data are provided by SoilWeb, ultimately sourced from from the current SSURGO snapshot. Data are returned as \code{raster} objects representing area proportion falling within 800m cells. Data are only available in CONUS and returned using an albers equal area / NAD83 corrdinate reference system.
}
\note{
This is a work in progress.
}
\examples{
\donttest{

if(requireNamespace("curl") &
   curl::has_internet()) {
  
  library(raster)
  library(rasterVis)
  library(viridis)
  library(soilDB)
  library(mapview)
  
  taxa <- 'vertisols'
  x <- taxaExtent(taxa, level = 'order')
  a <- aggregate(x, fact = 5)
  
  taxa <- 'ustalfs'
  x <- taxaExtent(taxa, level = 'suborder')
  a <- aggregate(x, fact = 5)
  
  taxa <- 'haplohumults'
  x <- taxaExtent(taxa, level = 'greatgroup')
  a <- aggregate(x, fact = 5)
  
  taxa <- 'Typic Haploxerepts'
  x <- taxaExtent(taxa, level = 'subgroup')
  a <- aggregate(x, fact = 5)
  
  
  levelplot(a, margin = FALSE, scales = list(draw = FALSE), col.regions = viridis, main = names(a))
  
  mapview(a, col.regions = viridis, na.color = NA, use.layer.names = TRUE)
  
  
}

}

}
\author{
D.E. Beaudette
}
