% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDB-package.R
\docType{data}
\name{us_ss_timeline}
\alias{us_ss_timeline}
\title{Timeline of US Published Soil Surveys}
\format{
A data.frame with 5209 observations on the following 5 variables.
\itemize{
\item \code{"ssa"}: Soil Survey name, a character vector
\item \code{"year"}: Year of publication, a numeric vector
\item \code{"pdf"}: Does a manuscript PDF document exist? a logical vector
\item \code{"state"}: State abbreviation, a character vector
}
}
\source{
https://www.nrcs.usda.gov/wps/portal/nrcs/soilsurvey/soils/survey/state/
}
\description{
This dataset contains the years of each US Soil Survey was published.
}
\details{
This data was web scraped from the NRCS Soils Website. The scraping
procedure and a example plot are included in the examples section below.
}
\keyword{datasets}
